new reducer organization
